got the image generator to now generate images from the slug. only problem is that difficulty is stuck on hard mode and keeps generating 3 word slugs
