+++ src/components/imageGenerator.jsx
@@ -1,32 +1,85 @@
 import { Configuration, OpenAIApi } from "openai";
 import { useState } from "react";
 import apiKey from "./API"
+import { generateSlug } from "random-word-slugs";
 
 function ImageGenerator(props) {
-  const [prompt, setPrompt] = useState(props.prompt);
   const configuration = new Configuration({
     apiKey: apiKey,
   });
-
   const openai = new OpenAIApi(configuration);
   const [result, setResult] = useState("");
 
-  const generateImage = async () => {
+  const generateImage = async (prompt) => {
     const res = await openai.createImage({
-      prompt: prompt, // use the prompt state value
+      prompt: prompt,
       n: 1,
       size: "512x512",
     });
-
     setResult(res.data.data[0].url);
   };
 
-  
+  const handleSelectDifficulty = (selectedDifficulty) => {
+    let slug;
+    if (selectedDifficulty === "easy") {
+      slug = generateSlug(1, {
+        format: "title",
+        partsOfSpeech: ["noun"],
+        categories: {
+          adjective: [],
+          noun: ["animals", "thing"],
+        },
+      });
+    } else if (selectedDifficulty === "medium") {
+      slug = generateSlug(2, {
+        format: "title",
+        partsOfSpeech: ["adjective", "noun"],
+        categories: {
+          adjective: ["color"],
+          noun: ["animals", "thing", "transportation"],
+        },
+      });
+    } else {
+      slug = generateSlug(3, {
+        format: "title",
+        partsOfSpeech: ["adjective", "adjective", "noun"],
+        categories: {
+          adjective: ["color", "size", "shape"],
+          noun: [
+            "animals",
+            "thing",
+            "transportation",
+            "people",
+            "profession",
+            "technology",
+          ],
+        },
+      });
+    }
+    console.log(slug)
+    generateImage(slug);
+  };
+
   return (
     <div>
-      <button id="generateBtn" onClick={generateImage} className="text-2xl border-2 border-solid border-zinc-900 p-px bg-gray-300">generate</button>
-      <div id="image-container" className="app-main text-2xl border-2 border-solid border-zinc-900 max-w-xl flex justify-around p-px bg-gray-300">
-      <img className="object-fill aspect-auto" src={result} alt="result" data-prompt={`${prompt}`} />
+      
+      <button
+        id="generateBtn"
+        onClick={() => handleSelectDifficulty(props.selectedDifficulty)}
+        className="text-2xl border-2 border-solid border-zinc-900 p-px bg-gray-300"
+      >
+        generate
+      </button>
+      <div
+        id="image-container"
+        className="app-main text-2xl border-2 border-solid border-zinc-900 max-w-xl flex justify-around p-px bg-gray-300"
+      >
+        <img
+          className="object-fill aspect-auto"
+          src={result}
+          alt="result"
+          data-prompt={props.prompt}
+        />
       </div>
     </div>
   );
